Map and Arrays for prefixes
Categorize tags and remove prefixes using Maps and Arrays.
A quick write-up, but works for the time being to get an idea for what is wanted.

Due to the data-tag-category's 'origin' being used for more than artist's names, prefixes are used to categorize the artist tag. There are also other prefixes such as 'editor' that are the same as 'artist'. This should be no problem as all artists' tags start with the 'artist' or 'editor' prefix.

Suffixes only applies when removing a prefix, this will probably be changed later to apply without removing the prefix.
It was coded to give an idea of wanting to change tagNames from "oc:originalcharacter" to "originalcharacter_(oc)".